requirements.txt
playwright==1.41.0
playwright-stealth==1.0.0
python-dotenv==1.0.0
aiofiles==23.2.1
rich==13.7.0
click==8.1.7
